Marge test to Master

+++ db/migrate/20200416151013_create_notifications.rb
@@ -2,7 +2,7 @@ class CreateNotifications < ActiveRecord::Migration[6.0]
   def change
     create_table :notifications do |t|
       t.bigint :user_id, null: false
-      t.integer :recipient_id, null: false
+      t.bigint :recipient_id, null: false
       t.datetime :read_at
       t.string :action, null: false
       t.string :notifiable_type
